changed square metres from int to double

+++ MadalinaBalaita/Imobile/Imobile.Core/Estates/Flat.cs
@@ -3,7 +3,7 @@ namespace Imobile.Core.Estates
 {
     public class Flat : Estate
     {   
-        public Flat(int sm1,string location1,string condition1) : base(sm1,location1,condition1)
+        public Flat(double sm1,string location1,string condition1) : base(sm1,location1,condition1)
         {
 
         }
