+++ MadalinaBalaita/Imobile/Imobile.Core/Estates/UrbanLand.cs
@@ -4,7 +4,7 @@ namespace Imobile.Core.Estates
 {
     public class UrbanLand:Estate
     {
-        public UrbanLand(int sm1, int cadastral_nr, string usage) : base(sm1, cadastral_nr,usage)
+        public UrbanLand(double sm1, int cadastral_nr, string usage) : base(sm1, cadastral_nr,usage)
         {
 
         }
